Filter more devices name and use kube-node group
The device name is now filtered removing the prefix '/dev/'.
If a '/' is still present in the device, we currently
raise an AssertionError, as ansible see only dm-x devices.

Also, the 'ansible_play_hosts' variable is used for the storage
check instead of iterating on all the hosts.
This way the target pf the playbook would be only the one we want to
validate the storage on

Issue #231